Set up Rock Paper Scissors project
Created a readme file with pseudocode for the project
Created a boilerplate HTML that links to the js file
Created a JS file.
Added a function to get a random number from the computer, and
match the number to either 'rock', 'paper' or 'scissors'.
Began a function for comparing the user's selection with the computer's.