Fix scoping for game-ending logic
Got rid of playToFive() and moved the logic that detects the score and
ends the game inside the playRound() function.

Also moved the playerWin() and playerLose() functions inside the body
of the playRound() function to keep things a bit cleaner.

Currently the win/lose message only fires on the first click *after*
one player reaches 5 points. That's next on the list.